move signature and ext to lrr lib
