+++ src/catapult/controller.py
@@ -7,9 +7,9 @@ from pathlib import Path
 from typing import Dict, List
 
 from catapult.connections import common
-from catapult.constants import ALLOWED_LRR_EXTENSIONS
 from catapult.cache import archive_hash_exists, insert_archive_hash
 from catapult.lanraragi.client import LRRClient
+from catapult.lanraragi.constants import ALLOWED_LRR_EXTENSIONS
 from catapult.lanraragi.utils import compute_sha1, compute_archive_id, get_signature_hex, is_valid_signature_hex
 from catapult.models import ArchiveMetadata, ArchiveUploadRequest, ArchiveUploadResponse, ArchiveValidateResponse, ArchiveValidateUploadStatus, MultiArchiveUploadResponse
 from catapult.utils import archive_contains_corrupted_image
